fix nodes joining cluster

+++ kube-startup.sh
@@ -21,11 +21,14 @@ curl -sSL https://raw.githubusercontent.com/coreos/flannel/v0.9.1/Documentation/
 # get slave local ips
 sudo ip -s -s neigh flush all
 sudo ping -b 10.0.0.255 -c 5 #replace with own ipaddress .255
-PI_IPS=$(arp -e | grep b8:27:eb | awk '{print $1}')
+PI_IPS=$(arp -e | grep eth0 | awk '{print $1}')
 
-for IP in PI_IPS; do
-    ssh pirate@$IP sudo $JOIN_COMMAND
-    ssh pirate@$IP sudo reboot
+for IP in $PI_IPS; do
+    status=$(ssh -o BatchMode=yes -o ConnectTimeout=5 $IP echo ok 2>&1)
+    if [[ $status == ok ]] ; then
+        ssh pirate@$IP sudo $JOIN_COMMAND
+        ssh pirate@$IP sudo reboot
+    fi
 done;
 
 cd ~/arm-kube-yarn; make;
